AppUser id changed to int

+++ src/main/java/app/controller/AdminController.java
@@ -28,7 +28,7 @@ public class AdminController {
     }
 
     @DeleteMapping("/delete")
-    public ResponseEntity<Long> callDeleteAppUser(@RequestParam long id) {
+    public ResponseEntity<Integer> callDeleteAppUser(@RequestParam int id) {
         appUserService.deleteAppUser(id);
         return new ResponseEntity<>(id, HttpStatus.OK);
